tests/wpt/tests/html/canvas/element/shadows/2d.shadow.composite.1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.shadow.composite.1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.shadow.composite.1</h1>
<p class="desc">Shadows are drawn using globalCompositeOperation</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Shadows are drawn using globalCompositeOperation");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.globalCompositeOperation = 'xor';
  ctx.shadowColor = '#f00';
  ctx.shadowOffsetX = 100;
  ctx.fillStyle = '#0f0';
  ctx.fillRect(-100, 0, 200, 50);

  _assertPixelApprox(canvas, 50,25, 0,255,0,255, 2);

});
</script>
